src/backend/requirements.txt
boto3==1.34.152
Flask==3.0.3
json5==0.9.5
pandas==2.2.2 
python-dotenv==1.0.1
scikit-learn==1.5.1
